refactor : FolderListPage로 폴더 정보 API 호출 책임 이관
- App.jsx 에서 폴더 정보 API 호출하는 구조에서 FolderListPage 에서
  호출하는걸로 변경
