+++ src/components/App.jsx
@@ -7,27 +7,11 @@ import { getFolders } from "../api/api";
 import Header from "./Header";
 
 function App() {
-  const [folders, setFolders] = useState([]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const folders = await getFolders();
-
-        if (!folders) return;
-        setFolders(folders);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-    fetchData();
-  }, []);
-
   return (
     <>
       <Header />
       <div className={styles.body}>
-        <FolderListPage folders={folders} />
+        <FolderListPage />
       </div>
       <Footer />
     </>
